HTTP Error handling for Apple IAP verification
Before this change, we were seeing the following errors:

    appstore: *json.SyntaxError: invalid character '<' looking for beginning of value

It was happening when Apple's API returned a 503 error with the body
"<html><body><b>Http/1.1 Service Unavailable</b></body> </html>".

This change returns a custom error class for status codes >= 400, which
also satisfies interface{ Temporary() bool } for status codes >= 500.
(This is inspired by errors in the standard library's net package, and
helps us decide whether to retry a failed request.)